Send normal termination metrics on wifi off
When turning wifi off, the interface gets torn down and empty
LinkProperties are received before wifi calls stop(). This causes a loss
of provisioning to be logged, instead of normal termination.

Watch interface link status up/down events, and when provisioning is
lost when the interface is down, consider it a normal termination.

Bug: 151796056
Test: manual: turn wifi off, observe events
Test: atest NetworkStackIntegrationTests (see also test-only change)
Change-Id: I9d086a199de0017aa425219d20882211423925e0