Implement DHCP IPv6-Only preferred option.
An IPv6-capable host includes this option in the Parameter Request
List option and willing to forgo obtaining an IPv4 address for a
specific assigned period(V6ONLY_WAIT). DHCP server responds with
DHCPOFFER or DHCPACK with this option if network can provide IPv6
connectivity.

Currently whether or not including IPv6-only preferred option in the
PRL is controlled by experiment flag.

Bug: 159670487
Test: atest NetworkStackTests NetworkStackIntegrationTests
Change-Id: I89653271197164a7b7f638c43cb3bf4d28817b9f