Apply jarjar rules to NetworkStaticLibTests
Each test target in NetworkStackCoverageTests needs to have appropriate
jarjar rules applied: this was not the case for NetworkStaticLibTests,
even though jarjar rules are already applied in NetworkStackTests and
NetworkStackIntegrationTests.

NetworkStaticLib is used by multiple modules that are each supposed to
jarjar it to their own package, so jarjar rules cannot be applied on the
test target itself.
Define an android library to apply the jarjar rules before including the
tests into NetworkStaticLibTests.

Bug: 152461494
Test: atest NetworkStackCoverageTests

Change-Id: I2afebd0e0fbfde2ed8c9daa396077761db23409f