Set default impl for INetworkMonitorCallbacks.
In S, calling an unimplemented AIDL method causes an exception when
there's no default implementation registered for the interface. The
change doesn't actually affect the network stack module because the
unimplemented methods are all oneway which doesn't trigger the exception
(because the clients cannot know if the call was successful or not).

However, that becomes a problem in the test where the binder objects are
mocked (or spy'ed) in the same process. In that case the exception
occurs because the clients (to be specific the generated proxy class)
can be notified with the unimplemented method.

Fixing the problem by registering the default implementation for the
problematic interface INetworkMonitorCallbacks.

Bug: 167632276
Test: mts/networking/gce-unittests-r-build-module-on-qt-platform
Change-Id: Ie21363062ee7066863dfd0d6a2d0d3a0d2812a0d