Add test for NetworkStackUtils.attachRaFilter
The test uses buildRaPacket, which was defined in
IpClientIntegrationTest and is moved to a utility class in
frameworks/libs/net.

It creates a socket, sends a ICMPv6 echo packet and verifies that the
packet is received by the socket. Then attaches the RA filter, sends
another ICMPv6 echo packet and an RA, and verifies that only the RA is
received.

Bug: 168868607
Test: atest NetworkStackCoverageTests:NetworkStackUtilsIntegrationTest
Change-Id: I067570aa2ea6964a940b6baa8e935c3f6676643e