Revert "Revert "Add new network quirk parcelable in the NetworkAttributes""
Add new network quirk parcelable in the NetworkAttributes.

This change is the first step to detect the network quirk happened on
a given network and save the associated information to the IPMS.

Bug: 157534516
Test: atest NetworkStackTests

This reverts commit_4210a7239c73357ddde5c0424f39c4af90e5df9b.

Reason for revert:
Roll forward the previous change, as the CtsStrictJavaPackagesTestCases
breakage has been addressed by adding the jarjar rule in wifi module.

Change-Id: I0351a8e544530c05c957352e29f6bc479f01197e