Add warning comments regarding downgrades
The current downgrade code may have a bug, and at least
is not covered by enough testing.
This is fine as long as the database version does not change
(it has been the same since Q), but the schema needs to be
frozen until downgrades are properly tested.

Downgrades were not expected to be necessary, but due to
a bug in at least Q and R, applications using device-encrypted
storage may not have their device-encrypted data directory
rolled back on package rollback, so the database downgrade
code would be run.

Bug: 171340630
Change-Id: Ice41f756a1824869b95cf509361f17d083f7ecbf
Test: m