Add NetworkStackTests / NetworkStackNextTests to mainline-presubmit
The tests are already in presubmit, but the general presubmit target is
not used in mainline-prod because there is no usable device image there.
Add them to mainline-presubmit so mainline-prod is also covered. The
test targets are unit tests that run in the order of 5min.

This involves moving the unit tests from device-tests to general-tests,
which is fine as they do not use device-specific signing keys.

Change-Id: I034a82b3c58859ee2dc364509549f31fb5c670b7
Test: only treehugger can validate the config change