Add min_sdk_version to NetworkStack dev bin/tests
This is necessary to run the NetworkStackNextTests on older platforms.
NetworkStackNextTests should pass on the same targets as
NetworkStackTests, so that when bumping NetworkStack SDK to
NetworkStackNext, the tests are known to pass.
The "next" targets should only differ by the build SDK version, not the
min_sdk_version, so there is no other difference to take care of when
bumping SDK.

Fixes: 173995979
Test: NetworkStackNextTests can be installed / run on released R

Change-Id: I10032ba6f2476fa3d56572fc507868658f86c70e