Do not run NetworkStackNextTests in mainline_presubmit
NetworkStackNextTests cannot run in mainline_presubmit because it targets development SDK, and mainline_presubmit is designed to run on released SDK. Setting min_sdk_version does not
resolve it, as the target SDK still mismatches.

NetworkStackTests is the appropriate test suite in that case. NetworkStackNextTests still runs in
classic presubmit, which is run on the latest development images.

Another alternative would be to have NetworkStackNextTests target a lower SDK version, but
this would remove coverage for tests on the "next" SDK, which may lead to surprises when
bumping SDK versions.

Change-Id: Ie720b382963ca8a6441d3bf5b32675080bf2ddf6
Test: TH configuration
Fixes: 173995979