Skip testNoFdLeaks in IpClientRootTest
The test looks at FDs from the current process, and uses a mock
ConnectivityManager, so it fails in IpClientRootTest.

The test should be fixed so it can count FDs from the NetworkStack
process, but remove it from IpClientRootTest so that
IpClientIntegrationTest can be added to presubmit and avoid this
kind of problems in the future.

Bug: 160541918
Change-Id: I435b33ddfdb057f1bea68b51dbb30a6c43378768
Test: Forrest run