Fix flaky testNotifyNetwork_WithforceReevaluation test
The test incorrectly starts two NetworkMonitors on single test.
But the callback object that is used for verifying the result is
the same one. Thus, the test will be flaky depending on the
timing on when verification starts and when the verification
result comes up. Update the test to use the same NM object.

Bug: 175330953
Test: atest NetworkMonitorTest#\
testNotifyNetwork_WithforceReevaluation --rerun-until-failure 50

Change-Id: I78463f83f0627783f2f49f7f13805ccac2df27d1