Cleanup testOverrideVendorId static member after running unittest.
"testOverrideVendorId" variable is just used in the DhcpPacket unit
test, we set it to "android-dhcp-???" to check whether the vendor id
sent in the packet is expected. However, in the code coverage test,
since it's static member, its value might affect the following test cases
which are checking "mVendorId" member(e.g. IpClientIntegrationTest),
depends on the execute order of test case. Hence, reset the
"testOverrideVendorId" variable after running unit test.

Bug: 169128961
Test: atest NetworkStackTests
Change-Id: If2a974cb7a9c1cbba57d49beef29736862b99232