Disable -current shims in mainline-prod.
The current shims will in general fail to build in mainline-prod
because they depend on APIs that do not exist there.

Disable them.

The Merged-In clause comes from the unrelated change
aosp/1605016. It is intended to ensure that this change does not
merge into any downstream branches.

Bug: 167645754
Test: m NetworkStack CtsNetTestCasesLatestSdk CtsHostsideNetworkTests CtsHostsideNetworkTestsApp
Change-Id: I9d37a261ac31c97c3a6e9d5153619ab30dc07eb2
Merged-In: I2ec0aa737804cc3ea5fce9326f84c1e4b5e4972c