DO NOT MERGE: Enable/disable current targets via soong var instead of inline.
Add a soong variable to enable/disable all targets that build
against the current SDK. This makes it easy to enable or disable
those targets all together. It also makes it much easier to
automerge Android.bp changes from AOSP to mainline-prod without
merge conflict, because the build target definitions are always
the same, only the variable at the top of the file changes.

Bug: 167645754
Test: m NetworkStackApiStableShims succeeds
Test: m NetworkStackApiCurrentShims fails with "unknown target" error
Ignore-AOSP-First: needed to solve merge conflicts
Change-Id: Ie0e7d4f73efcf1942aadf8b9d67e4de914307c09