Add shims for NetworkRequest
The setUids method in NetworkRequest.Builder is updated to take
a set of integer Range instead of a set of UidRange. This causes
the existing cts net tests targeting for stable APIs failed.
Thus, create NetworkRequestShim as a wrapper that can be used in
common code.

Ignore-AOSP-First: Prevent race in automerger

Bug: 172183305
Test: m
Change-Id: I0b79c73e82b6d0040cb3e5a822c40f62c329ce3f