Fix the failure testCustomizedDhcpOptions_disallowedOption.
This test intends to verify that a "disallowed" customized DHCP client
option shouldn't appear in DHCPDISCOVER/DHCPREQUEST packet, hostname
option might not be a good candidate for this case, because it's enabled
since S, device still can fetch the hostname from system settings and
includes it in DHCPDISCOVER even if that's a "disallowed" customized
option. Update the test with MTU option(ideally only DHCP Server should
assign MTU option, but it's just for testing here).

Bug: 185867658
Test: atest NetworkStackIntegrationTests
Change-Id: Iebbf0d6db3c888826db0471d1fb8c82751470167