Update the target SDK version to 30 in the NetworkStackIntegrationTest.
The default MAC address of loopback interface turns out anonymized MAC
when system app is running on the target SDK version < 30, to prevent
NPE crash (see aosp/1693685). That will break the testInterfaceParams
which is asserting the InterfaceParams#hasMacAddress is false based
on the null MAC address on loopback interface. Update using the target
SDK 30 to keep the same behavior as before.

Bug: 188468203
Test: atest NetworkStackIntegrationTests NetworkStackTests
Test: atest NetworkStackCoverageTests
Change-Id: I679ac029899fcb7c5b0b33df474ee11eb08a6f74