Do not apply jarjar rules on intermediate libs
Using jarjar on intermediate libraries causes issues when they are
combined with other libraries that use different jarjar rules, as the
resulting binary may contain multiple incompatible copies of the
classes. Instead, jarjar should be used on the final artifact, after
combining the libraries.

Ensure NetworkStackJarJarRules is not applied on intermediate libraries
like NetworkStackTestsLib, or NetworkStackIntegrationTestsLib, but
instead apply it on NetworkStackIntegrationTests or
NetworkStackCoverageTests.

Bug: 192535368
Test: atest TetheringCoverageTests TetheringIntegrationTests \
          CtsTetheringTest
Change-Id: Ifbbe7b2969db080e3daac240bd31297a6fd0150c