Only send skipped validation result to AIDL 11+
NetworkMonitor assumed that S would be able to handle
NETWORK_VALIDATION_RESULT_SKIPPED, but this is not the case on S builds
that do not have an update of the connectivity module.

Instead of checking for the platform being S, check for the actual
version of the remote via its stable AIDL callback version.

Bug: 197276728
Test: atest NetworkMonitorTest
Change-Id: Ib7e7ad2028414d9de402a0b780d17d03aac19247