Read URLs from handler thread
It's a preparation for dynamically changing the probing URLs
after NetworkMonitor being created. The probing URLs are final
members that are read at the constructor. The URLs may be
updated in the following commit. The URLs will be used in the
probing thread instead of handler thread which could cause
problems. Thus, take the probing URLs as parameters read from
handler thread to the probing thread.

Bug: 189738017
Bug: 181919957
Test: atest NetworkStackTests
Change-Id: Iee2d31a9cd37ad83792653cb796e304aff616d37