Add onIpv6AddressRemoved method in IpClientLinkObserver.callback.
Currently when IpClient receives onInterfaceAddressRemoved callback and
if the removed address is a global IPv6 address, remove it from the set
of gratuitous NA target address as well. To remove the dependency of the
overrided NetworkObserver usage in IpClient, move the code to the new
added callback onIpv6AddressRemoved, which is also triggered from the
IpClientLinkObserver when an IPv6 address was removed and keep the
calling order consistent as before.

Bug: 163492391
Test: atest NetworkStackTests NetworkStackIntegrationTests
Change-Id: I2a36fed6efb443a7c1dc13c76ec3fb83db2df437