Fix binding to TestNetworkStackService
TestNetworkStackServiceClient used the production INetworkStackConnector
action in the service component instead of the test action, causing it
to bind to the production service instead of the test service.

The two services are identical apart from access restrictions, so this
was not noticed while the integration tests are running with
MAINLINE_NETWORK_STACK permissions. But when trying to change the test
to use root permissions instead, binding would fail saying that the
permission is missing, as the test service is not being used.

Fix the component to use the TestNetworkStackService action.

Bug: 203175155
Test: atest NetworkStackRootTests with followup change
Change-Id: Ifeb83a3f028ac337b3a8fcff65e97ac040e8179e