Fix the probe wake lock duration when mcast_resolicit is enabled.
Fix the bug introduced in aosp/1890315, adding the number of resolicit
in setNeighborParameters is incorrect, because this function might be
called(e.g. setNeighbourParametersPostRoaming) by multiple times,
besides, where the number of resolicit is an invalid number(-1), that
results in an incorrect total number of probe solicits. For example,
when number of solicit is 5, then total number of probe solicits by
adding number of resolicit is 4(5-1). Instead, the more reasonable
approach would be adding the resolicit number iif mcast_resolicit
is enabled and getProbeWakeLockDuration is called.

Also fix the error of fail to set neighbor parameter due to the loss
of mocked number of solicit and interval in the integration test.

Bug: 162944199
Test: atest NetworkStackTests NetworkStackIntegrationTests
Change-Id: Ifa3a02218c5c77092214812f2a1f03ce76b34ae9