Deliver LinkProperties(with mParcelSensitiveFields=false) to wifi.
mParcelSensitiveFields is marked with transient keyword, it should be
always false in the IIpClientCallback caller's side, except on Go
devices where IpClient is running in the system_server, it's possible
that no parceling happens and the object is sent as true, which results
in wifi throws the UnsupportedOperationException on disconnection and
then reboot. To keep the consistent behavior, deliver LinkProperties
with mParcelSensitiveFields=false to wifi upon callback is triggered.

Bug: 203741634
Test: m
Test: connect to wifi AP and verify the wifi works.
Test: atest testDhcpClientCaptivePortalApiEnabled_ParcelSensitiveFields
Change-Id: I91ab70aea9032a06a640f568d8548d7d6f53ba0d