Only adopt MANAGE_TEST_NETWORKS, not all shell permissions.
IpClientIntegrationTestCommon#setupTapInterface adopts the shell
permission identity in order to create the tap interface. This
causes the test to lose all non-shell permissions, and because
the test runs as the production networkstack user, it causes the
production networkstack on the device to lose non-shell
permissions as well.

Instead, only adopt the MANAGE_TEST_NETWORKS permission.

Fix: 215307963
Test: atest NetworkStackNextIntegrationTests:IpClientIntegrationTest
Test: no system crash if NetworkMonitor validates while test is running
Change-Id: Ib6c1d642549a2940664e63b374ed33d6fb00caa1