Remove redundant network verification for as-is network
If the network is unvalidated and the user selects "use as is",
the device will go to EvaluatingPrivateDnsState and then go to
ValidatedState eventually if everything is OK.

Then when the CMD_FORCE_REEVALUATION comes, the device will go
to EvaluatingState and handles CMD_REEVALUATE. Before, the
device will go to ProbingState to probe, the probe is redundant
because the user has selected "use as is" and the result is
always valid. Instead, the device should go to
EvaluatingPrivateDnsState straightly and then go to ValidatedState
eventually, like what the device do when the user selects
"use as is".

Also update the test accordingly. The NetworkMonitor state
should transit to ValidatedState eventually without a
redirected URL.

Bug: 216428888

Signed-off-by: zhujiatai <zhujiatai@xiaomi.com>
Change-Id: I14a403ae670392d1b8763756d69594353821d3d0