Fix : recent Connectivity module with old NetworkStack module
The guarantees that the NetworkStack module will be as recent
as the Connectivity module are a little bit weak, so we should
make sure the rare configuration where Connectivity is more
recent should work.

In the absence of a satisfactory solution which :
- Doesn't cause log spew
- Is easy to manage process-wise
- Is scalable
This implements a simple solution where clients of the
network-stack-client lib only calls the new
notifyNetworkConnectedParcel method in T+. This is safe because
a T device can never revert to a version of the NetworkStack
that doesn't support this, and is in fact optimal since before
T the additional info is never useful (the one relevant field
in NetworkAgentConfig is always false, for lack of a public API
to make it true in S-).

Test: ConnectivityServiceTest
      NetworkMonitorTest
Change-Id: Ieaade098c55bc296d593c90e518de727b639d527