Delete NetworkStackHostTests from postsubmit.
NetworkStackHostTests has failed for a while since the test infra
was updated with InstallUtilsHost, the initilization should be moved
into the setUp, otherwise, the null check for the TestInformation
instance will fail when calling initilization#getTestFile(). However,
it looks like there is still other failure for this test. Consider
that module release team will verify the module install and rollback
testing, probably we could delete this test from postsubmit completely
and keep the test dashboard healthy.

Bug: 202288248
Test: m
Change-Id: Iaec6217539e41d760e40f45031eab7484665b1bf