Bump NetworkStack to target SDK 33
Bump NetworkStack to become NetworkStackNext, and create api34 shims
that are only included in the new NetworkStackNext.

This is a no-op on devices currently in the field as they use SDK 31
(which was the previous target) or 32, and there are no behavior changes
affecting NetworkStack in 32:
developer.android.com/about/versions/12/12L/summary#get-apps-ready

NetworkStackNext was also used until now by development API 33 builds,
so they will see no change either.

Ignore-AOSP-First: API 33 prebuilts not yet in AOSP. Will cherry-pick
                   with minor modifications.
Bug: 216396304
Test: atest NetworkStackTests NetworkStackNextTests
Change-Id: Ibb12df8a54e373e08c07195f94d0def9cb022ad5