Use module_current for NetworkStack stable shims
In branches that do not yet have a prebuilt module_33 SDK dropped, use
"module_current" instead of "module_33" in the build rules.

This change should not be merged into branches that have the prebuilt
SDK, as per the merged-in.

Bug: 216396304
Test: m
Merged-In: Ibb12df8a54e373e08c07195f94d0def9cb022ad5
Change-Id: I51c9158a364b6a0b960c411843a4ee10e8b473fe