Enable the debug log by default in the IpClientLinkObserver.
Add logError implementation in NetlinkMonitor, otherwise, if any
execption happens when reading packets, that won't be logged.

Turn on the debug log by default in the IpClientLinkObserver, to
better understand what happened exactly from the bug report, and
only log the address/route/DNS info when the LinkProperties changes,
that should help reduce the verbose log as much as possible.

finally, also log the terrible failure when IpClient is trying to set
the interface parameters but NetlinkMonitor isn't running yet at this
ponit.

Bug: 163492391
Bug: 232680956
Bug: 233191218
Test: atest NetworkStackIntegrationTests
Test: manually test, turn on/off wifi multiple times to see if wifi
      still works.
Change-Id: I87070a684496c5eeebaf8208f275ab1a90d30495