Create DhcpPacketLib
The library can be used by CTS to parse/generate DHCP packets, rather
than include the whole NetworkStack android library (which includes its
manifest and many classes that are not needed).

Also move the test URL configuration constants to NetworkStackConstants
so that the tests can still reference them.

This change removes the UnsupportedAppUsage annotation as it is not
imported by the library. That annotation is actually not useful anymore
as DhcpResults is not in the bootclasspath already so it cannot be used
by apps (since change I356ee7b24613e84b8a90ddccd6d25d0742d13ae7 in
frameworks/base).

Test: atest CtsNetTestCases
Change-Id: Ifbf6435c50b1cf6c91785c81f5d81bcf9f430220