Add log.wtf when attempting to send EVENT_LINKADDRESS_CONFIGURED to DHCP Client.
DHCP client becomes null in IpClient#RunningState due to some reason and
NPE is thrown then when attempting to send a message to DHCP client,
it's observed in some stress testings from OEM, but not sure what the
root cause is. Add log.wtf when this case does happen, this is useful to
see how often this case occurs on fields and also give us the trace log
for further debugging.

Bug: 203174383
Test: atest NetworkStackIntegrationTests
Change-Id: I78029b8a7ff0a4e0a2f63f1ef0c40e6fbe15cb82